requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# AI & ML Libraries
huggingface-hub==0.34.4
transformers==4.55.2
torch==2.8.0
tokenizers==0.21.4
safetensors==0.6.2
accelerate==1.10.0

# FastAPI Dependencies
python-multipart==0.0.6
pydantic==2.11.7
pydantic-core==2.33.2

# File & Environment Handling
python-dotenv==1.0.0
aiofiles==23.2.1
Pillow==10.0.1

# HTTP & Networking
requests==2.31.0
httptools==0.6.4
websockets==15.0.1
urllib3==2.5.0
certifi==2025.8.3

# Utilities
tqdm==4.67.1
click==8.2.1
colorama==0.4.6
packaging==25.0
filelock==3.18.0
fsspec==2025.7.0
watchfiles==1.1.0

# Data Processing
numpy==2.2.6
PyYAML==6.0.2
regex==2025.7.34

# Template & Markup
Jinja2==3.1.6
MarkupSafe==3.0.2

# Math & Symbolic Computing
sympy==1.14.0
mpmath==1.3.0
networkx==3.4.2

# System & Type Checking
psutil==7.0.0
typing-extensions==4.14.1
typing-inspection==0.4.1
annotated-types==0.7.0

# Async & HTTP
anyio==3.7.1
sniffio==1.3.1
h11==0.16.0
idna==3.10
charset-normalizer==3.4.3

# Compatibility
importlib-metadata==8.7.0
zipp==3.23.0
exceptiongroup==1.3.0